Modified to include custom id which auto increments.

+++ SplitListWebApi/ApiFormat/ShadowTables/PantryItem.cs
@@ -5,9 +5,9 @@ namespace ApiFormat.ShadowTables
 {
     public class PantryItem
     {
-        public double PantryModelID { get; set; }
+        public int PantryModelID { get; set; }
         public PantryModel PantryModel { get; set; }
-        public double ItemModelID { get; set; }
+        public int ItemModelID { get; set; }
         public ItemModel ItemModel { get; set; }
         public int Amount { get; set; }
     }
