+++ SplitListWebApi/ApiFormat/ShadowTables/RecipeItem.cs
@@ -5,10 +5,10 @@ namespace ApiFormat.ShadowTables
 {
     public class RecipeItem
     {
-        public double ItemModelID { get; set; }
+        public int ItemModelID { get; set; }
         public ItemModel ItemModel { get; set; }
 
-        public double RecipeModelID { get; set; }
+        public int RecipeModelID { get; set; }
         public RecipeModel RecipeModel { get; set; }
     }
 }
