+++ SplitListWebApi/ApiFormat/ShadowTables/ShoppingListItem.cs
@@ -5,10 +5,10 @@ namespace ApiFormat.ShadowTables
 {
     public class ShoppingListItem
     {
-        public double ShoppingListModelID { get; set; }
+        public int ShoppingListModelID { get; set; }
         public ShoppingListModel ShoppingListModel { get; set; }
 
-        public double ItemModelID { get; set; }
+        public int ItemModelID { get; set; }
         public ItemModel ItemModel { get; set; }
 
         public int Amount { get; set; }
